# ENH: Make deferring connections the default for main window again for better performance