--- a/pydm/data_plugins/__init__.py
+++ b/pydm/data_plugins/__init__.py
@@ -29,7 +29,19 @@ __plugins_initialized = False
 
 
 @contextmanager
-def connection_queue(defer_connections=False):
+def connection_queue(defer_connections: bool = False) -> None:
+    """
+    Creates a queue for holding channel connections and potentially processing them at a later time. When
+    defer_connections is set to True, the exit from the context manager will not result in any connections being made.
+    This allows for a more responsive end user experience on displays with many connections as it will load without
+    needing to establish all connections first.
+    Parameters
+    ----------
+    defer_connections : bool
+        Whether or not to defer making the actual connections to channels at a later time. Note that if this
+        is set to true, a call to establish_queued_connections() must be made at some point later as the queue
+        will not be automatically be processed by the context manager in this case.
+    """
     global __CONNECTION_QUEUE__
     global __DEFER_CONNECTIONS__
     if __CONNECTION_QUEUE__ is None:
@@ -41,7 +53,11 @@ def connection_queue(defer_connections=False):
     establish_queued_connections()
 
 
-def establish_queued_connections():
+def establish_queued_connections() -> None:
+    """
+    Processes all channels in the deferred connection queue establishing the actual connection for each. Upon
+    completion, will reset the global connection queue to None.
+    """
     global __DEFER_CONNECTIONS__
     global __CONNECTION_QUEUE__
     if __CONNECTION_QUEUE__ is None: