@ZLLentz 2023-02-22
I don't have suggestions about the code beyond Ken's suggestions, but I ran this with the `pmps-ui` with live PVs.

This causes the screen to open and be ready for user input in ~10-15 seconds, and then it slowly fills in all the PVs properly. Working as intended with great results :+1:. Previously, user input was delayed up to 30+ seconds.

It breaks my filtering on one of the tabs- my gui is convinced that all the widgets on that tab are staying disconnected, despite showing values, so it wants to hide them all by default. I'll check a bit to see what the discrepancy is.