@ZLLentz 2023-02-22
As far as I can tell, any `PyDMChannel` objects I create and call `connect` on in code prior to the queue being emptied never actually get connected. I don't understand why this happening after some looking through both the pydm source and the pmps-ui source.