@jbellister-slac 2023-03-31
I have not forgotten about this one, just pushing it off to the next release for now. The issue with the tab in the pmps display not loading properly turned out to be the use of native python threading as opposed to the pyqt version. I did not go deep into the pyqt code to figure out why yet, but it's likely signal/slot connections not working properly. Switching over to a pyqt thread makes it work fine.

I still did want to take another pass at improving performance though as well as testing thoroughly again before putting this back up for review.